+++ CHANGELOG.md
@@ -5,12 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-05-10
+
+- remove react-query
+- remove tx-provider
+- improve loading indicators
+- improve chain spec loading and type with JSON parsing
+- update next to 15.3.0
+- add turbopack to `dev` script
+
 ## [1.2.0] - 2025-04-10
 
 ### Added
 
 - Light client with webworkers support for Polkadot chains
-- Create 2 providers for either `rpc-api-provider` or `lightclient-api-provider`
+- Create 2 providers
+  - `lightclient-api-provider` (default)
+  - `rpc-api-provider` for rpc connections (faster but more centralized)
 - Support for parachain connections (polkadot assetHub, paseo asset hub)
 - add a CHANGELOG.md